Modernise CUDA resampling functions
 - Ditch old texture objects and use up-to-date ones
 - Make texture objects managed
 - Ditch CUDA symbols and pass them as kernel function parameters

+++ reg-lib/cuda/_reg_resampling_gpu.h
@@ -21,7 +21,7 @@ void reg_resampleImage_gpu(nifti_image *sourceImage,
                            cudaArray *sourceImageArray_d,
                            float4 *positionFieldImageArray_d,
                            int *mask_d,
-                           int activeVoxelNumber,
+                           size_t activeVoxelNumber,
                            float paddingValue);
 
 extern "C++"
@@ -29,5 +29,5 @@ void reg_getImageGradient_gpu(nifti_image *sourceImage,
                               cudaArray *sourceImageArray_d,
                               float4 *positionFieldImageArray_d,
                               float4 *resultGradientArray_d,
-                              int activeVoxelNumber,
+                              size_t activeVoxelNumber,
                               float paddingValue);
